+++ sync/startbuttons.py
@@ -1,57 +1,63 @@
-#!/usr/bin/python3
-
-import time
-from time import sleep
+from gpiozero import LED, Button
+from signal import pause
 import os
-import RPi.GPIO as GPIO
+import time
+
+#this is the new botton rigger script since the "old" one has trouble with debouncing
+
+continue_running=True
+
+#define the bouncetime here
+bounce_time = 3
+
+#led colors
+led_blau  = LED(11)
+led_rot   = LED(10)
+led_gruen = LED(9)
+
+#the main 3 functions of the buttons:
+
+BTN1 = ("sudo /var/www/sync/clockdisplay &")
+BTN2 = ("sudo /var/www/sync/clockdisplay &")
+BTN3 = ("sudo /var/www/sync/clockdisplay &")
+
+#what to trigger:
+
+def das_Erste():
+	led_blau.on()
+	os.system(BTN1)
+	print("we are doing the first thing")
+	time.sleep(bounce_time)
+	led_blau.off()
+
+def das_Zweite():
+	led_rot.on()
+	os.system(BTN2)
+	print("we are doing the second thing")
+	time.sleep(bounce_time)
+	led_rot.off()
+
+def das_Dritte():
+	led_gruen.on()
+	os.system(BTN3)
+	print("we are doing the third thing")
+	time.sleep(bounce_time)
+	led_gruen.off()
+
+#hauptprogramm:
+def main():
+	global continue_running
+	button1=Button(27, pull_up=False)
+	button1.when_pressed = das_Erste
+
+	button2=Button(18, pull_up=False)
+	button2.when_pressed = das_Zweite
+
+	button3=Button(17, pull_up=False)
+	button3.when_pressed = das_Dritte
+
+	while(continue_running):
+          time.sleep(0.1)
 
 
-#the functions for the buttons
-
-BTN1 = ("sudo /var/www/sync/startimage &")
-BTN2 = ("sudo /var/www/sync/startimage &")
-BTN3 = ("sudo /var/www/sync/startimage &")
-
-
-#here we define each button function:
-
-def button1_callback(channel):
-   os.system ("sudo /usr/bin/python /var/www/sync/ledblue.py &")
-   os.system (BTN1)
-   print("Button1 was pushed!")
-
-def button2_callback(channel):
-    os.system ("sudo /usr/bin/python /var/www/sync/ledred.py &")
-    os.system (BTN2)
-    print("Button2 was pushed!")
-
-def button3_callback(channel):
-    os.system ("sudo /usr/bin/python /var/www/sync/ledgreen.py &")
-    os.system (BTN3)
-    print("Button3 was pushed!")
-
-
-#intialize the buttons as inputs
-
-GPIO.setwarnings(False) # Ignore warning for now
-GPIO.setmode (GPIO.BCM) #BCM uses the GPIO numbers while BOARD uses the actual pin number
-GPIO.setup(17, GPIO.IN, pull_up_down=GPIO.PUD_DOWN) # Set gpio17 button1 to be an input pin and set initial value to be pulled low (off)
-GPIO.setup(18, GPIO.IN, pull_up_down=GPIO.PUD_DOWN) # Set gpio18 button2 to be an input pin and set initial value to be pulled low (off)
-GPIO.setup(27, GPIO.IN, pull_up_down=GPIO.PUD_DOWN) # Set gpio27 button3 to be an input pin and set initial value to be pulled low (off)
-
-#the main programm
-class MainProgram():
-	GPIO.add_event_detect(27,GPIO.RISING,callback=button1_callback,bouncetime=4000) # Setup event on pin 17 rising edge
-	GPIO.add_event_detect(18,GPIO.RISING,callback=button2_callback,bouncetime=4000) # Setup event on pin 18 rising edge
-	GPIO.add_event_detect(17,GPIO.RISING,callback=button3_callback,bouncetime=4000) # Setup event on pin 27 rising edge
-
-print("running")
-
-#this would kill the programm with an enter
-#message = input("Press enter to quit/n/n") # Run until someone presses enter
-
-#keeps the programm running forever:
-while(True):
-   MainProgram()
-
-GPIO.cleanup() # Clean up
+if __name__ == "__main__": main()